requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
httpx==0.27.0
python-dotenv==1.0.1
google-generativeai==0.3.2
nltk==3.8.1
spacy==3.7.2
transformers==4.40.1
torch==2.2.2
newspaper3k==0.2.8
lxml==5.2.1
chromadb==0.4.24
tqdm==4.66.1
requests==2.31.0
regex==2023.12.25
streamlit==1.32.2
sentence-transformers==2.6.1
huggingface_hub==0.23.0
